Refcator helpers
