+++ src/index.ts
@@ -4,7 +4,7 @@ import chalk from 'chalk'
 import { Command, Option } from 'commander'
 import { CloudProvider, Output, OutputFormats } from './constants'
 import CloudChiprCliProvider from './cli/cloud-chipr-cli-provider'
-import { LoggerHelper } from './helpers'
+import { logFile } from './helpers'
 require('dotenv').config()
 
 const command = new Command()
@@ -35,7 +35,7 @@ cloudChiprCli
 try {
   command.parseAsync(process.argv).catch(e => {
     const filename = `./.c8r/logs/${(new Date()).toISOString().slice(0, 10)}.log`
-    LoggerHelper.logFile(filename, e.message, e)
+    logFile(filename, e.message, e)
     if (command.getOptionValue('verbose') === true) {
       console.error(chalk.red(chalk.underline('Error:'), `Failed on executing command due to: ${e.message}. \nThe trace log can be found in ${filename} directory.`))
       console.error(chalk.red(e.stack))
@@ -45,6 +45,6 @@ try {
   })
 } catch (e) {
   const filename = `./.c8r/logs/${(new Date()).toISOString().slice(0, 10)}.log`
-  LoggerHelper.logFile(filename, e.message, e)
+  logFile(filename, e.message, e)
   console.error(chalk.red(chalk.underline('Error:'), `Failed on executing command due to: ${e.message}. \nThe trace log can be found in ${filename} directory.`))
 }
